third_party/WebKit/LayoutTests/sensor/idl-Gyroscope.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  // Test that Gyroscope interface exists
  assert_true('Gyroscope' in window);
}, 'Test that the Gyroscope interface is present.');
</script>
